backend/requirements.txt
# Django and REST Framework
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Celery and Redis
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0

# Stock Data
yfinance>=0.2.36,<1.0
pandas>=2.0,<3.0
numpy>=1.24,<2.0

# Utilities
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
